changed list objects in content to hashmaps
Former-commit-id: da4efbfc5a363c9a1fcff6dac2515754940e7b1d

+++ Domain_Objects/src/main/java/potatoes/project/domain_objects/User.java
@@ -52,7 +52,7 @@ public class User {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long userID;
+    private int userID;
     
     public User() {
     	
@@ -85,7 +85,7 @@ public class User {
         return password;
     }
     
-    public long getUserID() {
+    public int getUserID() {
     	return userID;
     }
